pyproject.toml
[tool.poetry]
name = "rag-knowledge-api"
version = "1.0.0"
description = "RAG-powered Knowledge Base API with FAISS"
authors = ["Roy305 <royantbryant@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
llama-index = "^0.9.0"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.3.1"
openai = "^1.10.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
bcrypt = "4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
numpy = "^2.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
ruff = "^0.1.11"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.poetry.scripts]
test = "pytest"
test-cov = "pytest --cov=app --cov-report=html"